Move copy_from_guest to provision/base.py
... taking cmd_mkcp with it and extracting quote() to tmt.utils.
